requirements.txt
amqp==5.2.0
anyascii==0.3.2
appdirs==1.4.4
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.1.0
auth==0.5.3
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.11.2
billiard==4.2.0
binaryornot==0.4.4
blinker==1.7.0
Brotli==1.1.0
calmjs.parse==1.3.1
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
confluent-kafka==2.3.0
constantly==23.10.4
cookiecutter==2.5.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==41.0.7
csscompressor==0.9.5
cssselect2==0.7.0
daphne==4.0.0
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.8
django-adminactions==2.3.0
django-allauth==0.59.0
django-appconf==1.0.6
django-baton==2.9.1
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-classy-tags==4.1.0
django-cms==3.11.4
django-compressor==4.4
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.5
django-formtools==2.5.1
django-guardian==2.4.0
django-haystack==3.2.1
django-imagekit==5.0.0
django-import-export==3.3.5
django-model-utils==4.3.1
django-modelcluster==6.1
django-payments==2.0.0
django-paypal==2.0
django-permissionedforms==0.1
django-phonenumber-field==7.3.0
django-redis==5.4.0
django-redis-cache==3.0.1
django-rest-jwt==0.1.46
django-rest-swagger==2.2.0
django-sekizai==4.1.0
django-storages==1.14.2
django-stubs==4.2.7
django-stubs-ext==4.2.7
django-tables2==2.7.0
django-taggit==4.0.0
django-tenants==3.6.1
django-timezone-field==6.1.0
django-tinymce4-lite==1.8.0
django-treebeard==4.7
djangocms-admin-style==3.2.6
djangorestframework==3.14.0
dnspython==2.5.0
draftjs-exporter==2.1.7
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
eventlet==0.34.3
falcon==3.1.3
filetype==1.2.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
freetype-py==2.3.0
greenlet==3.0.3
gunicorn==21.2.0
html5lib==1.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsmin==3.0.1
kafka-python==2.0.2
kombu==5.3.4
l18n==2021.3
lxml==5.0.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.3
mdurl==0.1.2
mongoengine==0.27.0
mypy==1.7.1
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
oauthlib==3.2.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.2
phonenumberslite==8.13.29
pilkit==3.0
Pillow==10.1.0
pillow-heif==0.14.0
ply==3.11
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.25.1
pycparser==2.21
Pygments==2.17.2
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==1.7.1
pymongo==4.6.1
PyMySQL==1.1.0
pyOpenSSL==23.3.0
pypdf==4.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
rcssmin==1.1.1
redis==3.5.3
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rjsmin==1.2.1
rlPyCairo==0.3.0
sentry-sdk==1.39.1
service-identity==23.1.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.27
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
telepath==0.3.1
text-unidecode==1.3
tinycss2==1.2.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
types-python-dateutil==2.8.19.14
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.1.0
vine==5.1.0
wagtail==5.2.2
wcwidth==0.2.12
webencodings==0.5.1
Werkzeug==3.0.1
Willow==1.6.3
xhtml2pdf==0.2.14
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.1
